Move the HTTP recorder to its own file
